@@ -48,22 +48,23 @@
     "eslint-plugin-jsx-a11y": "^6.2.1",
     "eslint-plugin-react": "^7.13.0",
     "express": "^4.17.1",
-    "karma": "^6.3.14",
+    "karma": "^6.4.4",
     "karma-chrome-launcher": "^2.2.0",
     "karma-mocha": "^1.3.0",
-    "karma-webpack": "^4.0.2",
+    "karma-webpack": "^5.0.1",
     "mocha": "^6.1.4",
     "pretty-error": "^2.0.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "webpack": "^5.94.0",
-    "webpack-cli": "^3.3.4",
+    "webpack": "^5.104.1",
+    "webpack-cli": "^6.0.1",
     "webpack-dev-middleware": "^5.3.4"
   },
   "dependencies": {
+    "@twemoji/api": "17.0.2",
+    "@types/react-dom": "^19.2.3",
     "lodash.isequal": "^4.5.0",
     "prop-types": "^15.7.2",
-    "@twemoji/api": "17.0.2"
+    "react": "^19.2.3",
+    "react-dom": "^19.2.3"
   },
   "peerDependencies": {
     "react": ">=16.4.2"
